HPCC-22542 Activity Widget sending hundreds of requests to WUInfo
During a change to correct the progress graphic that was out of sync in the WUDetailsWidget a hasCompleted event fired in multiple places. This caused
monitoring(timer) to start on the Activity widget needlessly. The Activity widget has its own timer that can monitor its un-completed workunits. I have added
an indentifier that destinguishes it from the WUDetailsWidget.

Signed-off: Miguel Vazquez <miguel.vazquez@lexisnexisrisk.com>